use permissions from API and remove constants reference
Issue: AAH-2016
