+++ CHANGES/2016.task
@@ -1,0 +1,1 @@
+Use API permissions in lieu of constants permissions in editable role form.
